requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.0
py-clob-client==0.18.0
python-dotenv==1.0.1
cryptography==43.0.1
eth-account==0.13.0
web3==7.2.0
PyJWT==2.9.0
pyotp==2.9.0
qrcode==7.4.2
